[Update] Kodgen to latest version + [WIP] Namespace registration

+++ Refureku/Generator/Submodules/Kodgen/Include/Parsing/EParsingError.h
@@ -1,3 +1,10 @@
+/**
+*	Copyright (c) 2020 Julien SOYSOUVANH - All Rights Reserved
+*
+*	This file is part of the Kodgen library project which is released under the MIT License.
+*	See the README.md file for full license details.
+*/
+
 #pragma once
 
 #include <string>
@@ -6,7 +13,7 @@
 
 namespace kodgen
 {
-	enum class EParsingError : uint16
+	enum class EParsingError : uint8
 	{
 		/**
 		*	The wrong macro has been used to mark an entity,
